Do not hardcode the 'root' group name, use the 'gid' for that. Non-Linux systems, i.e. *BSD might not have a group 'root', but have group wheel. What is in common everywhere is the 'gid' of these groups is '0'.
Setting the group => '0' instead of group => 'root' is a common
pattern seen in many other modules.

This fixes the setting of the group for the npmrc file.
For consistency, I also changed the group in manifests/repo/nodesource/yum.pp
even it doesn't affect me.

Ran into the problem on OpenBSD.