WIP: Validate soft dependencies exist in params.pp
As a potential way of helping the user with soft dependencies, we can
attempt to load the module metadata. If we don't get any metadata back,
that probably means the module is not installed. We can then warn the
user that they need to install it.

We can also inspect returned metadata for version and author
compatibility, though ideally we would be very very lax.