Fix idempotence of unquoted secrets (#390)
* Fix idempotence of unquoted secrets

The fix in https://github.com/voxpupuli/puppet-nodejs/pull/386 fixes
idempotence for secrets that contain characters (like '=') that cause
NPM to quote them in npmrc. Secrets that end up unquoted in npmrc are
not handled correctly however. This fix makes the quotes optional when
checking if the config value is already set correctly.